Pulse waves working a little better

+++ src/main/java/com/rosscon/llce/components/audio/NES/DMC.java
@@ -8,7 +8,7 @@ public class DMC extends Channel{
     }
 
     @Override
-    public double getSample() {
+    public int getSample() {
         return 0;
     }
 }
