+++ src/main/java/com/rosscon/llce/components/audio/NES/Noise.java
@@ -7,7 +7,7 @@ public class Noise extends Channel {
     }
 
     @Override
-    public double getSample() {
+    public int getSample() {
         return 0;
     }
 }
