+++ src/main/java/com/rosscon/llce/components/audio/NES/Triangle.java
@@ -7,7 +7,7 @@ public class Triangle extends Channel{
     }
 
     @Override
-    public double getSample() {
+    public int getSample() {
         return 0;
     }
 }
